+++ tests/test_path_extraction.rs
@@ -326,6 +326,7 @@ mod svg_conversion_tests {
             artifact_type: None,
             reading_order: None,
             layer: None,
+            xobject_path: None,
         }
     }
 
@@ -392,6 +393,7 @@ mod svg_conversion_tests {
             artifact_type: None,
             reading_order: None,
             layer: None,
+            xobject_path: None,
         };
 
         // Generate curve path data
@@ -425,6 +427,7 @@ mod svg_conversion_tests {
             artifact_type: None,
             reading_order: None,
             layer: None,
+            xobject_path: None,
         };
 
         // Rectangle should be converted to M L L L Z
@@ -453,6 +456,7 @@ mod svg_conversion_tests {
             artifact_type: None,
             reading_order: None,
             layer: None,
+            xobject_path: None,
         };
 
         assert_eq!(round_cap.line_cap, LineCap::Round);
@@ -477,6 +481,7 @@ mod svg_conversion_tests {
             artifact_type: None,
             reading_order: None,
             layer: None,
+            xobject_path: None,
         };
 
         assert_eq!(bevel_join.line_join, LineJoin::Bevel);